requirements.txt
PySide6>=6.7.0
ultralytics>=8.3.0
opencv-python>=4.10.0.84
numpy>=1.26.0
Pillow>=10.3.0
torch>=2.2.0
pyinstaller>=6.8.0
